+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Mark all Modrinth mods as required when exporting unless explicitely marked as unsupported
+- Modrinth mrpack downloads not showing size on instance installer
 
 ### Misc