+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.15.2
+## 3.4.16.0
 
 ### New Features
+- Add setting for user to select a custom download path for browser downloaded mods
 
 ### Fixes
 - Mark all Modrinth mods as required when exporting unless explicitely marked as unsupported